completed refactoring sidebar logic and interfaces
