+++ frontend/package.json
@@ -12,7 +12,7 @@
     "@emotion/react": "^11.13.3",
     "@emotion/styled": "^11.13.0",
     "@mui/icons-material": "^6.1.3",
-    "@mui/material": "^6.1.3",
+    "@mui/material": "^6.1.4",
     "d3": "^7.9.0",
     "d3-hexbin": "^0.2.2",
     "leaflet": "^1.9.4",
