+++ frontend/src/utils/interfaces.ts
@@ -1,0 +1,24 @@
+
+export type Province = {
+  Aimag: string;
+  Year: number;
+  Camel: number;
+  Cattle: number;
+  Goat: number;
+  Horse: number;
+  Sheep: number;
+};
+
+export type Provinces = {
+  Aimag: string;
+  Year: number[];
+  Camel: number[];
+  Cattle: number[];
+  Goat: number[];
+  Horse: number[];
+  Sheep: number[];
+};
+
+export type Mongolia = {
+  [Aimag: string]: Province[];
+};
